pgfplot bar patterns; evaluation corrected partly

+++ umlModels/ActivityExamples/AcademicProblems/triangleClassificator/TriangleClassificator.h
@@ -18,7 +18,7 @@ extern "C"{
 #endif
 
 int classify(const int* const x, const int* const y, const int* const z);
-float classifyNoLogic(const float* const x, const float* const y, const float* const z);
+double classifyNoLogic(const double* const x, const double* const y, const double* const z);
 
 #ifdef __cplusplus
 }
